NO-SNOW: jdbc - extract native lib loading into NativeLibraryLoader
Addresses review feedback: JNICoreTransport's responsibility is handling
messages over JNI, not finding and loading the native library. Move that
logic to a dedicated NativeLibraryLoader class in the same package.
